+++ GameEngine/Code/Visual/OpenGLRenderer.h
@@ -47,9 +47,9 @@ namespace Engine::Visual
             std::vector<Material> materials;
             glm::mat4 worldMatrix;
 
-			int GetVertexCount() const override
+			size_t GetVertexCount() const override
 			{
-				return static_cast<int>(vertices.size());
+				return vertices.size();
 			}
         };
