+++ GameEngine/Code/Visual/VulkanRenderer.h
@@ -101,9 +101,9 @@ namespace Engine::Visual
             VkBuffer uniformBuffer;
             VkDeviceMemory uniformBufferMemory;
 
-			int GetVertexCount() const override 
+			size_t GetVertexCount() const override 
             {
-				return static_cast<int>(vertices.size());
+				return vertices.size();
 			}
         };
